Own the visibility helpers in uiSpecification
uiSpecification/utils imported visibleFieldSet from
databaseEngine/completion, the only import in that direction in the
package. FieldVisibilityMap describes currentlyVisibleMap's output, so
both now live beside their producer. The status walk also hand-rolled
the component-parameters.hidden test that getFieldsMatchingCondition
applies behind includeStaticallyHidden; one exported
isFieldStaticallyHidden now serves both sites.
